Goodbye Instagram, hello new backend api
New API will be very similar to Instagram api, just for convenience of
the front end.

Collection of “data” with location property containing URI.
Pagination object with the nextMaxId - this is the smallest id in the
collection of objects. Effectively paging backwards through the
collection, based on the last id the client saw.

This isn’t how the flickr api actually works, it has stateful numbered
pagination but this is problematic as items are easily lost/missed.
Instead we will map to this more generic approach on the server.